refactor:서비스 리팩토링 및 FavoriteTargetType 제거
- rankingService에서 FavoriteTargetType 관련 코드 삭제
- favoriteService로 기능 이전 및 구조 개선
- 불필요한 favorite.ts 파일 삭제
