--- src/constants/favorite.ts
@@ -1,6 +1,0 @@
-export enum FavoriteTargetType {
-  QUESTION = "QUESTION",
-  ANSWER = "ANSWER",
-  COMMENT = "COMMENT",
-  POST = "POST",
-}
