+++ src/services/rankingService.ts
@@ -1,21 +1,8 @@
 import prisma from "../lib/prisma.js";
-import {
-  Question,
-  Answer,
-  Comment,
-  CommunityPost,
-  Prisma,
-} from "@prisma/client";
-import { FavoriteTargetType } from "../constants/favorite.js";
 
 const rankingService = {
   getLikesCountRankings,
   getAnswerCountRankings,
-  getFavoriteCountByTargetType,
-  updateFavoriteCount,
-  getFavoriteCount,
-  incrementFavoriteCount,
-  decrementFavoriteCount,
 };
 
 export interface RankingUser {
@@ -190,134 +177,6 @@ async function getAnswerCountRankings(
       totalAnswerCount: user._count.answers,
     };
   });
-}
-
-async function getFavoriteCountByTargetType(
-  targetType: FavoriteTargetType,
-  targetId: number
-): Promise<number> {
-  const count = await prisma.favorite.count({
-    where: {
-      targetType,
-      targetId,
-    },
-  });
-  return count;
-}
-
-async function updateFavoriteCount(
-  targetType: FavoriteTargetType,
-  targetId: number
-): Promise<void> {
-  const count = await getFavoriteCountByTargetType(targetType, targetId);
-  await updateEntityFavoriteCount(targetType, targetId, {
-    favoriteCount: count,
-  });
-}
-
-async function getFavoriteCount(
-  targetType: FavoriteTargetType,
-  targetId: number
-): Promise<number> {
-  const result = await findUniqueEntity(targetType, targetId, {
-    select: { favoriteCount: true },
-  });
-  return result?.favoriteCount ?? 0;
-}
-
-// 좋아요 증가 캐시
-async function incrementFavoriteCount(
-  targetType: FavoriteTargetType,
-  targetId: number,
-  tx?: Prisma.TransactionClient
-): Promise<void> {
-  await updateEntityFavoriteCount(
-    targetType,
-    targetId,
-    {
-      favoriteCount: { increment: 1 },
-    },
-    tx
-  );
-}
-
-// 좋아요 감소 캐시
-async function decrementFavoriteCount(
-  targetType: FavoriteTargetType,
-  targetId: number,
-  tx?: Prisma.TransactionClient
-): Promise<void> {
-  await updateEntityFavoriteCount(
-    targetType,
-    targetId,
-    {
-      favoriteCount: { decrement: 1 },
-    },
-    tx
-  );
-}
-
-async function updateEntityFavoriteCount(
-  targetType: FavoriteTargetType,
-  targetId: number,
-  data: {
-    favoriteCount: number | { increment: number } | { decrement: number };
-  },
-  tx?: Prisma.TransactionClient
-): Promise<void> {
-  const prismaClient = tx || prisma;
-  switch (targetType) {
-    case FavoriteTargetType.QUESTION:
-      await prismaClient.question.update({ where: { id: targetId }, data });
-      break;
-    case FavoriteTargetType.ANSWER:
-      await prismaClient.answer.update({ where: { id: targetId }, data });
-      break;
-    case FavoriteTargetType.COMMENT:
-      await prismaClient.comment.update({ where: { id: targetId }, data });
-      break;
-    case FavoriteTargetType.POST:
-      await prismaClient.communityPost.update({
-        where: { id: targetId },
-        data,
-      });
-      break;
-    default:
-      throw new Error("Invalid target type");
-  }
-}
-
-async function findUniqueEntity<
-  T extends Question | Answer | Comment | CommunityPost
->(
-  targetType: FavoriteTargetType,
-  targetId: number,
-  select: { select: { favoriteCount: true } }
-): Promise<(T & { favoriteCount: number }) | null> {
-  switch (targetType) {
-    case FavoriteTargetType.QUESTION:
-      return prisma.question.findUnique({
-        where: { id: targetId },
-        ...select,
-      }) as Promise<(T & { favoriteCount: number }) | null>;
-    case FavoriteTargetType.ANSWER:
-      return prisma.answer.findUnique({
-        where: { id: targetId },
-        ...select,
-      }) as Promise<(T & { favoriteCount: number }) | null>;
-    case FavoriteTargetType.COMMENT:
-      return prisma.comment.findUnique({
-        where: { id: targetId },
-        ...select,
-      }) as Promise<(T & { favoriteCount: number }) | null>;
-    case FavoriteTargetType.POST:
-      return prisma.communityPost.findUnique({
-        where: { id: targetId },
-        ...select,
-      }) as Promise<(T & { favoriteCount: number }) | null>;
-    default:
-      throw new Error("Invalid target type");
-  }
 }
 
 export default rankingService;
